Continue prettification using fixed-width integer aliases

+++ kernel/src/math.h
@@ -2,15 +2,15 @@
 #define LENSOR_OS_MATH_H
 
 struct Vector2 {
-	unsigned int x;
-	unsigned int y;
+	u64 x;
+	u64 y;
 
 	Vector2() {
 		x = 0;
 		y = 0;
 	}
 
-	Vector2(unsigned int _x, unsigned int _y) {
+	Vector2(u64 _x, u64 _y) {
 		x = _x;
 		y = _y;
 	}
